Dashboard, Aprovação de Respostas

+++ src/pages/pesquisa/index.tsx
@@ -28,23 +28,47 @@ export default function pesquisa(props: interfProps) {
 
     const [pesquisas, setpesquisas] = useState<Array<interfpesquisa>>([]);
 
-    function deleteResearch(id: number) {
-        api.delete(`/pesquisas/${id}`, {
-            headers: {
-                Authorization: "Bearer " + props.token,
-            },
-        })
-            .then((res) => {
-                findPesquisa();
-                Swal.fire(
-                    'Deletado com Sucesso!',
-                    'Click em OK!',
-                    'error'
-                  )
-            })
-            .catch((erro) => {
-                console.log(erro);
+    function deleteResearch(id: number, total_respostas: number) {
+        if (total_respostas > 0) {
+            Swal.fire({
+                title: 'Erro!',
+                text: 'Não é possível deletar uma pesquisa com respostas!',
+                icon: 'error',
+                confirmButtonText: 'Ok'
             });
+            router.push("/pesquisa");
+        } else {
+            Swal.fire({
+                title: 'Deletar Pesquisa?',
+                text: "Você não poderá reverter isso!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Sim, deletar!'
+            }).then((result) => {
+                if (result.value) {
+
+                    api.delete(`/pesquisas/${id}`, {
+                        headers: {
+                            Authorization: "Bearer " + props.token,
+                        },
+                    })
+                        .then((res) => {
+                            findPesquisa();
+                            Swal.fire(
+                                'Deletado com Sucesso!',
+                                'Click em OK!',
+                                'error'
+                            )
+                        })
+                        .catch((erro) => {
+                            console.log(erro);
+                        });
+                }
+            });
+        }
+
     }
 
     function findPesquisa() {
@@ -54,11 +78,10 @@ export default function pesquisa(props: interfProps) {
             },
         })
             .then((res) => {
-                if(res.data.status === "Token is Expired"){
+                if (res.data.status === "Token is Expired") {
                     //Adicionar Mensagem de Login Expirado
-                    alert("Token is Expired");
                     Swal.fire({
-                        title: 'Token is Expired',
+                        title: 'Token Expirado!',
                         text: '',
                         icon: 'error',
                         confirmButtonText: 'OK'
@@ -67,7 +90,7 @@ export default function pesquisa(props: interfProps) {
                     }
                     );
                 } else {
-                setpesquisas(res.data);
+                    setpesquisas(res.data);
                 }
             })
             .catch((erro) => {
@@ -75,18 +98,19 @@ export default function pesquisa(props: interfProps) {
             });
     }
 
-    function getStatus(status: boolean){
-        if(status){
-            return <small className="badge bg-success rounded-pill"><BsCheck/></small>
+    function getStatus(status: string) {
+        console.log(status);
+        if (status === '1') {
+            return <small className="badge bg-success rounded-pill"><BsCheck /></small>
         } else {
-            return <small className="badge bg-secondary rounded-pill"><BsXLg/></small>
+            return <small className="badge bg-secondary rounded-pill"><BsXLg /></small>
         }
     }
 
     useEffect(() => {
         findPesquisa();
     }, []);
-    return(
+    return (
         <>
             <Head>
                 <title>Pesquisa</title>
@@ -100,46 +124,55 @@ export default function pesquisa(props: interfProps) {
                     <div
                         className="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-4 pb-2 mb-3 border-bottom"
                     >
-                        <h2><BsFillPersonFill/> Pesquisas</h2>
+                        <h2><BsFillPersonFill /> Pesquisas</h2>
                         <div
                             className="btn-toolbar mb-2 mb-md-0"
                         >
                             <button type="button" onClick={() => router.push('/pesquisa/novo')}
-                            className="btn btn-success rounded-pill"><BsPlusLg/> Adicionar</button>
+                                className="btn btn-success rounded-pill"><BsPlusLg /> Adicionar</button>
                         </div>
                     </div>
                 </>
                 <div className="row">
-                {pesquisas.map((pesquisa: interfpesquisa) => (
-                   <div className="col-sm-4 p-2" key={pesquisa.id}>
-                     <div className="card border border-primary shadow-lg w-100 h-100">
-                       <div className="card-body">
-                         <h5 className="card-title">
-                            <div className="row">
-                                <div className="col-12 col-md-8 col-sm-12 col-lg-10">
-                                    {pesquisa.tema}
-                                </div>
-                                <div className="col-12 col-md-3 col-sm-12 col-lg-2">
-                                    {getStatus(pesquisa.status)}
+                    {pesquisas.map((pesquisa: interfpesquisa) => (
+                        <div className="col-sm-4 p-2" key={pesquisa.id}>
+                            <div className="card border border-primary shadow-lg w-100 h-100">
+                                <div className="card-body">
+                                    <h5 className="card-title">
+                                        <div className="row">
+                                            <div className="col-12 col-md-8 col-sm-12 col-lg-10">
+                                                {pesquisa.tema}
+                                            </div>
+                                            <div className="col-12 col-md-3 col-sm-12 col-lg-2">
+                                                {getStatus(pesquisa.status.toString())}
+                                            </div>
+                                        </div>
+                                    </h5>
+                                    <p className="card-text">{pesquisa.descricao}</p>
+                                    <div className="text-center">
+                                        <button type="button" onClick={() => router.push(`/pesquisa/respostas/${pesquisa.id}`)}
+                                            className="btn btn-primary btn-sm m-1">{pesquisa.total_respostas} Respostas</button>
+
+                                        <button type="button" className="btn btn-success btn-sm m-1"
+                                            onClick={() => {
+                                                router.push(`/pesquisa/${pesquisa.id}`)
+                                            }
+                                            }
+                                        ><BsPencil /> Editar</button>
+
+                                        <button
+                                            className="btn btn-danger btn-sm m-1"
+                                            onClick={() => {
+                                                deleteResearch(pesquisa.id, pesquisa.total_respostas);
+                                            }}
+                                        >
+                                            <BsTrash /> Apagar
+                                        </button>
+                                    </div>
                                 </div>
                             </div>
-                         </h5>
-                         <p className="card-text">{pesquisa.descricao}</p>
-                         <div className="text-center">
-                         <button type="button" onClick={() => router.push(`/pesquisa/respostas/${pesquisa.id}`)}
-                            className="btn btn-primary btn-sm m-1">{pesquisa.total_respostas} Respostas</button>
-
-                            <button type="button" className="btn btn-success btn-sm m-1"
-                            onClick={() => {
-                                router.push(`/pesquisa/${pesquisa.id}`)
-                            }
-                            }
-                            ><BsPencil/> Editar</button>
-                                </div>
-                       </div>
-                     </div>
-                   </div>
-                ))}
+                        </div>
+                    ))}
                 </div>
             </Menu>
         </>
@@ -148,7 +181,7 @@ export default function pesquisa(props: interfProps) {
 
 export const getServerSideProps: GetServerSideProps = async (contexto) => {
 
-    const {'painel-token': token} = parseCookies(contexto);
+    const { 'painel-token': token } = parseCookies(contexto);
 
     // console.log(token)
